remove base class as it introduce variable shadowing
