+++ src/main/java/com/redhat/sast/api/model/Job.java
@@ -9,7 +9,7 @@ import jakarta.persistence.*;
 
 @Entity
 @Table(name = "job")
-public class Job extends JobBase {
+public class Job {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -155,5 +155,109 @@ public class Job extends JobBase {
 
     public void setJobSettings(JobSettings jobSettings) {
         this.jobSettings = jobSettings;
+    }
+
+    public String getProjectName() {
+        return projectName;
+    }
+
+    public void setProjectName(String projectName) {
+        this.projectName = projectName;
+    }
+
+    public String getProjectVersion() {
+        return projectVersion;
+    }
+
+    public void setProjectVersion(String projectVersion) {
+        this.projectVersion = projectVersion;
+    }
+
+    public String getPackageName() {
+        return packageName;
+    }
+
+    public void setPackageName(String packageName) {
+        this.packageName = packageName;
+    }
+
+    public String getPackageNvr() {
+        return packageNvr;
+    }
+
+    public void setPackageNvr(String packageNvr) {
+        this.packageNvr = packageNvr;
+    }
+
+    public String getOshScanId() {
+        return oshScanId;
+    }
+
+    public void setOshScanId(String oshScanId) {
+        this.oshScanId = oshScanId;
+    }
+
+    public String getJiraLink() {
+        return jiraLink;
+    }
+
+    public void setJiraLink(String jiraLink) {
+        this.jiraLink = jiraLink;
+    }
+
+    public String getHostname() {
+        return hostname;
+    }
+
+    public void setHostname(String hostname) {
+        this.hostname = hostname;
+    }
+
+    public JobStatus getStatus() {
+        return status;
+    }
+
+    public void setStatus(JobStatus status) {
+        this.status = status;
+    }
+
+    public LocalDateTime getCreatedAt() {
+        return createdAt;
+    }
+
+    public void setCreatedAt(LocalDateTime createdAt) {
+        this.createdAt = createdAt;
+    }
+
+    public LocalDateTime getStartedAt() {
+        return startedAt;
+    }
+
+    public void setStartedAt(LocalDateTime startedAt) {
+        this.startedAt = startedAt;
+    }
+
+    public LocalDateTime getCompletedAt() {
+        return completedAt;
+    }
+
+    public void setCompletedAt(LocalDateTime completedAt) {
+        this.completedAt = completedAt;
+    }
+
+    public LocalDateTime getCancelledAt() {
+        return cancelledAt;
+    }
+
+    public void setCancelledAt(LocalDateTime cancelledAt) {
+        this.cancelledAt = cancelledAt;
+    }
+
+    public String getTektonUrl() {
+        return tektonUrl;
+    }
+
+    public void setTektonUrl(String tektonUrl) {
+        this.tektonUrl = tektonUrl;
     }
 }
